+++ packages/svelte-query/tests/createMutation/createMutation.svelte.test.ts
@@ -1,4 +1,4 @@
-import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { flushSync } from 'svelte'
 import { fireEvent, render } from '@testing-library/svelte'
 import { QueryClient } from '@tanstack/query-core'
@@ -22,7 +22,7 @@ describe('createMutation', () => {
     vi.useRealTimers()
   })
 
-  test('should be able to reset `error`', async () => {
+  it('should be able to reset `error`', async () => {
     const rendered = render(Reset, {
       props: { queryClient },
     })
@@ -38,7 +38,7 @@ describe('createMutation', () => {
     expect(rendered.getByText('Error: undefined')).toBeInTheDocument()
   })
 
-  test('should be able to call `onSuccess` and `onSettled` after each successful mutate', async () => {
+  it('should be able to call `onSuccess` and `onSettled` after each successful mutate', async () => {
     const onSuccessMock = vi.fn()
     const onSettledMock = vi.fn()
 
@@ -69,7 +69,7 @@ describe('createMutation', () => {
     expect(onSettledMock).toHaveBeenNthCalledWith(3, 3)
   })
 
-  test('should set correct values for `failureReason` and `failureCount` on multiple mutate calls', async () => {
+  it('should set correct values for `failureReason` and `failureCount` on multiple mutate calls', async () => {
     type Value = { count: number }
 
     const mutationFn = vi.fn<(value: Value) => Promise<Value>>()
@@ -108,7 +108,7 @@ describe('createMutation', () => {
     expect(rendered.getByText('Failure Reason: undefined')).toBeInTheDocument()
   })
 
-  test(
+  it(
     'should recreate observer when queryClient changes',
     withEffectRoot(async () => {
       const queryClient1 = new QueryClient()