+++ packages/solid-query-devtools/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/solid-query-devtools
 
+## 5.100.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.100.1
+  - @tanstack/solid-query@5.100.1
+
+## 5.100.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-devtools@5.100.0
+  - @tanstack/solid-query@5.100.0
+
 ## 5.99.2
 
 ### Patch Changes